+++ wp-content/plugins/nginx-helper/admin/class-purger.php
@@ -2,17 +2,15 @@
 /**
  * The admin-specific functionality of the plugin.
  *
- * @package    nginx-helper
+ * @package    gridpane-nginx-helper
  */
 
 /**
  * Description of purger
  *
- * @package    nginx-helper
+ * @package    gridpane-nginx-helper
  *
  * @subpackage nginx-helper/admin
- *
- * @author     rtCamp
  */
 abstract class Purger {
 
@@ -86,6 +84,19 @@ abstract class Purger {
 
 		$_post_id    = $comment->comment_post_ID;
 		$_comment_id = $comment->comment_ID;
+		$_post_type  = get_post_type( $_post_id );
+
+		$exclude_post_types = apply_filters( 'rt_nginx_helper_comment_change_exclude_post_types', array() );
+
+		if ( in_array( $_post_type, $exclude_post_types, true ) ) {
+			if ( 'nav_menu_item' !== $_post_type ) {
+				$this->log( '* * * * *' );
+				$this->log('* Post Type comment update - ' . $_post_type . ' - purge trigger excluded...');
+				$this->log('* Filter: rt_nginx_helper_comment_change_exclude_post_types');
+				$this->log( '* * * * *' );
+			}
+			return;
+		}
 
 		$this->log( '* * * * *' );
 		$this->log( '* Blog :: ' . addslashes( get_bloginfo( 'name' ) ) . ' ( ' . $blog_id . ' ). ' );
@@ -134,6 +145,8 @@ abstract class Purger {
 			return;
 		}
 
+		$_post_type = get_post_type( $post_id );
+
 		switch ( current_filter() ) {
 
 			case 'publish_post':
@@ -157,7 +170,6 @@ abstract class Purger {
 				break;
 
 			default:
-				$_post_type = get_post_type( $post_id );
 				$this->log( '* * * * *' );
 				$this->log( '* Blog :: ' . addslashes( get_bloginfo( 'name' ) ) . ' ( ' . $blog_id . ' ).' );
 				$this->log( "* Custom post type '" . $_post_type . "' :: " . get_the_title( $post_id ) . ' ( ' . $post_id . ' ).' );
@@ -169,8 +181,22 @@ abstract class Purger {
 
 		$this->log( 'Function purge_post BEGIN ===' );
 
+		# Here we need something more granular
+		# to allow a post type to trigger its purge
+		# but skip the homepage purge
+		# can use array, if found, skip purge homepage
+		# [ 'post_type1', 'post_type1' ]
+		# we can also do exceptions by post/page slug too
+
 		if ( 1 === (int) $nginx_helper_admin->options['purge_homepage_on_edit'] ) {
-			$this->_purge_homepage();
+			if ( ! in_array( $_post_type,  $nginx_helper_admin->options['homepage_purge_post_type_exceptions'], true ) ) {
+				$this->_purge_homepage();
+			} else {
+				$this->log('* * * * *');
+				$this->log('* Post Type update - ' . $_post_type . ' - homepage purge trigger excluded...');
+				$this->log('* ' . $_post_type . ' in constant array: RT_WP_NGINX_HELPER_HOMEPAGE_PURGE_EXCEPTIONS');
+				$this->log('* * * * *');
+			}
 		}
 
 		if ( 'comment_post' === current_filter() || 'wp_set_comment_status' === current_filter() ) {
@@ -232,7 +258,7 @@ abstract class Purger {
 				$url = get_sample_permalink( $post_id );
 
 				if ( ! empty( $url[0] ) && ! empty( $url[1] ) ) {
-					$url = str_replace( '%postname%', $url[1], $url[0] );
+					$url = str_replace( array('%postname%', '%pagename%'), $url[1], $url[0] );
 				} else {
 					$url = '';
 				}
@@ -263,7 +289,9 @@ abstract class Purger {
 
 			}
 
-			if ( 'post' === $_post_type ) {
+			$post_types = get_post_types( array( 'public' => true ) );
+
+			if ( in_array( $_post_type, $post_types, true ) ) {
 
 				$this->log( 'Purging date' );
 
@@ -387,7 +415,8 @@ abstract class Purger {
 		}
 
 		// Build a hash of the URL.
-		$hash = md5( $url_data['scheme'] . 'GET' . $url_data['host'] . $url_data['path'] );
+		$url_path = isset( $url_data['path'] ) ? $url_data['path'] : '';
+		$hash = md5( $url_data['scheme'] . 'GET' . $url_data['host'] . $url_path );
 
 		// Ensure trailing slash.
 		$cache_path = RT_WP_NGINX_HELPER_CACHE_PATH;
@@ -400,10 +429,12 @@ abstract class Purger {
 		 * Filters the cached file name.
 		 *
 		 * @since 2.1.0
+		 * @since 2.2.3 Purge URL argument `$url` were added.
 		 *
 		 * @param string $cached_file Cached file name.
+		 * @param string $url         URL to be purged.
 		 */
-		$cached_file = apply_filters( 'rt_nginx_helper_purge_cached_file', $cached_file );
+		$cached_file = apply_filters( 'rt_nginx_helper_purge_cached_file', $cached_file, $url );
 
 		// Verify cached file exists.
 		if ( ! file_exists( $cached_file ) ) {
@@ -524,7 +555,7 @@ abstract class Purger {
 
 		global $nginx_helper_admin;
 
-		if ( ! $nginx_helper_admin->options['enable_log'] ) {
+		if ( ! $nginx_helper_admin->is_nginx_log_enabled() ) {
 			return;
 		}
 
@@ -557,7 +588,7 @@ abstract class Purger {
 
 		global $nginx_helper_admin;
 
-		if ( ! $nginx_helper_admin->options['enable_log'] ) {
+		if ( ! $nginx_helper_admin->is_nginx_log_enabled() ) {
 			return;
 		}
 
@@ -683,12 +714,24 @@ abstract class Purger {
 		if ( function_exists( 'icl_get_home_url' ) ) {
 
 			$homepage_url = trailingslashit( icl_get_home_url() );
-			$this->log( sprintf( __( 'Purging homepage (WPML) ', 'nginx-helper' ) . '%s', $homepage_url ) );
+			$this->log(
+				sprintf(
+					/* translators: %s homepage URL */
+					__( 'Purging homepage (WPML) %s', 'gridpane-nginx-helper' ),
+					$homepage_url
+				)
+			);
 
 		} else {
 
 			$homepage_url = trailingslashit( home_url() );
-			$this->log( sprintf( __( 'Purging homepage ', 'nginx-helper' ) . '%s', $homepage_url ) );
+			$this->log( 
+				sprintf(
+					/* translators: %s homepage URL */
+					__( 'Purging homepage %s', 'gridpane-nginx-helper' ),
+					$homepage_url
+				)
+			);
 
 		}
 
@@ -698,29 +741,6 @@ abstract class Purger {
 
 	}
 
-	/**
-	 * Purge personal urls.
-	 *
-	 * @return bool
-	 */
-	private function _purge_personal_urls() {
-
-		global $nginx_helper_admin;
-
-		$this->log( __( 'Purging personal urls', 'nginx-helper' ) );
-
-		if ( isset( $nginx_helper_admin->options['purgeable_url']['urls'] ) ) {
-
-			foreach ( $nginx_helper_admin->options['purgeable_url']['urls'] as $url ) {
-				$this->purge_url( $url, false );
-			}
-		} else {
-			$this->log( '- ' . __( 'No personal urls available', 'nginx-helper' ) );
-		}
-
-		return true;
-
-	}
 
 	/**
 	 * Purge post categories.
@@ -731,7 +751,7 @@ abstract class Purger {
 	 */
 	private function _purge_post_categories( $_post_id ) {
 
-		$this->log( __( 'Purging category archives', 'nginx-helper' ) );
+		$this->log( __( 'Purging category archives', 'gridpane-nginx-helper' ) );
 
 		$categories = wp_get_post_categories( $_post_id );
 
@@ -740,7 +760,7 @@ abstract class Purger {
 			foreach ( $categories as $category_id ) {
 
 				// translators: %d: Category ID.
-				$this->log( sprintf( __( "Purging category '%d'", 'nginx-helper' ), $category_id ) );
+				$this->log( sprintf( __( "Purging category '%d'", 'gridpane-nginx-helper' ), $category_id ) );
 				$this->purge_url( get_category_link( $category_id ) );
 
 			}
@@ -758,7 +778,7 @@ abstract class Purger {
 	 */
 	private function _purge_post_tags( $_post_id ) {
 
-		$this->log( __( 'Purging tags archives', 'nginx-helper' ) );
+		$this->log( __( 'Purging tags archives', 'gridpane-nginx-helper' ) );
 
 		$tags = get_the_tags( $_post_id );
 
@@ -766,7 +786,7 @@ abstract class Purger {
 
 			foreach ( $tags as $tag ) {
 
-				$this->log( sprintf( __( "Purging tag '%1\$s' ( id %2\$d )", 'nginx-helper' ), $tag->name, $tag->term_id ) );
+				$this->log( sprintf( __( "Purging tag '%1\$s' ( id %2\$d )", 'gridpane-nginx-helper' ), $tag->name, $tag->term_id ) );
 				$this->purge_url( get_tag_link( $tag->term_id ) );
 
 			}
@@ -785,7 +805,7 @@ abstract class Purger {
 	 */
 	private function _purge_post_custom_taxa( $_post_id ) {
 
-		$this->log( __( 'Purging post custom taxonomies related', 'nginx-helper' ) );
+		$this->log( __( 'Purging post custom taxonomies related', 'gridpane-nginx-helper' ) );
 
 		$custom_taxonomies = get_taxonomies(
 			array(
@@ -799,7 +819,7 @@ abstract class Purger {
 			foreach ( $custom_taxonomies as $taxon ) {
 
 				// translators: %s: Post taxonomy name.
-				$this->log( sprintf( '+ ' . __( "Purging custom taxonomy '%s'", 'nginx-helper' ), $taxon ) );
+				$this->log( sprintf( '+ ' . __( "Purging custom taxonomy '%s'", 'gridpane-nginx-helper' ), $taxon ) );
 
 				if ( ! in_array( $taxon, array( 'category', 'post_tag', 'link_category' ), true ) ) {
 
@@ -813,11 +833,11 @@ abstract class Purger {
 					}
 				} else {
 					// translators: %s: Post taxonomy name.
-					$this->log( sprintf( '- ' . __( "Your built-in taxonomy '%s' has param '_builtin' set to false.", 'nginx-helper' ), $taxon ), 'WARNING' );
+					$this->log( sprintf( '- ' . __( "Your built-in taxonomy '%s' has param '_builtin' set to false.", 'gridpane-nginx-helper' ), $taxon ), 'WARNING' );
 				}
 			}
 		} else {
-			$this->log( '- ' . __( 'No custom taxonomies', 'nginx-helper' ) );
+			$this->log( '- ' . __( 'No custom taxonomies', 'gridpane-nginx-helper' ) );
 		}
 
 		return true;
@@ -830,7 +850,7 @@ abstract class Purger {
 	 */
 	private function _purge_all_categories() {
 
-		$this->log( __( 'Purging all categories', 'nginx-helper' ) );
+		$this->log( __( 'Purging all categories', 'gridpane-nginx-helper' ) );
 
 		$_categories = get_categories();
 
@@ -838,13 +858,13 @@ abstract class Purger {
 
 			foreach ( $_categories as $c ) {
 
-				$this->log( sprintf( __( "Purging category '%1\$s' ( id %2\$d )", 'nginx-helper' ), $c->name, $c->term_id ) );
+				$this->log( sprintf( __( "Purging category '%1\$s' ( id %2\$d )", 'gridpane-nginx-helper' ), $c->name, $c->term_id ) );
 				$this->purge_url( get_category_link( $c->term_id ) );
 
 			}
 		} else {
 
-			$this->log( __( 'No categories archives', 'nginx-helper' ) );
+			$this->log( __( 'No categories archives', 'gridpane-nginx-helper' ) );
 
 		}
 
@@ -858,7 +878,7 @@ abstract class Purger {
 	 */
 	private function _purge_all_posttags() {
 
-		$this->log( __( 'Purging all tags', 'nginx-helper' ) );
+		$this->log( __( 'Purging all tags', 'gridpane-nginx-helper' ) );
 
 		$_posttags = get_tags();
 
@@ -866,12 +886,12 @@ abstract class Purger {
 
 			foreach ( $_posttags as $t ) {
 
-				$this->log( sprintf( __( "Purging tag '%1\$s' ( id %2\$d )", 'nginx-helper' ), $t->name, $t->term_id ) );
+				$this->log( sprintf( __( "Purging tag '%1\$s' ( id %2\$d )", 'gridpane-nginx-helper' ), $t->name, $t->term_id ) );
 				$this->purge_url( get_tag_link( $t->term_id ) );
 
 			}
 		} else {
-			$this->log( __( 'No tags archives', 'nginx-helper' ) );
+			$this->log( __( 'No tags archives', 'gridpane-nginx-helper' ) );
 		}
 
 		return true;
@@ -885,7 +905,7 @@ abstract class Purger {
 	 */
 	private function _purge_all_customtaxa() {
 
-		$this->log( __( 'Purging all custom taxonomies', 'nginx-helper' ) );
+		$this->log( __( 'Purging all custom taxonomies', 'gridpane-nginx-helper' ) );
 
 		$custom_taxonomies = get_taxonomies(
 			array(
@@ -899,7 +919,7 @@ abstract class Purger {
 			foreach ( $custom_taxonomies as $taxon ) {
 
 				// translators: %s: Taxonomy name.
-				$this->log( sprintf( '+ ' . __( "Purging custom taxonomy '%s'", 'nginx-helper' ), $taxon ) );
+				$this->log( sprintf( '+ ' . __( "Purging custom taxonomy '%s'", 'gridpane-nginx-helper' ), $taxon ) );
 
 				if ( ! in_array( $taxon, array( 'category', 'post_tag', 'link_category' ), true ) ) {
 
@@ -915,11 +935,11 @@ abstract class Purger {
 					}
 				} else {
 					// translators: %s: Taxonomy name.
-					$this->log( sprintf( '- ' . esc_html__( "Your built-in taxonomy '%s' has param '_builtin' set to false.", 'nginx-helper' ), $taxon ), 'WARNING' );
+					$this->log( sprintf( '- ' . esc_html__( "Your built-in taxonomy '%s' has param '_builtin' set to false.", 'gridpane-nginx-helper' ), $taxon ), 'WARNING' );
 				}
 			}
 		} else {
-			$this->log( '- ' . __( 'No custom taxonomies', 'nginx-helper' ) );
+			$this->log( '- ' . __( 'No custom taxonomies', 'gridpane-nginx-helper' ) );
 		}
 
 		return true;
@@ -946,7 +966,7 @@ abstract class Purger {
 	 */
 	private function _purge_all_posts() {
 
-		$this->log( __( 'Purging all posts, pages and custom post types.', 'nginx-helper' ) );
+		$this->log( __( 'Purging all posts, pages and custom post types.', 'gridpane-nginx-helper' ) );
 
 		$args = array(
 			'posts_per_page' => 0,
@@ -961,12 +981,12 @@ abstract class Purger {
 
 			foreach ( $_posts as $p ) {
 
-				$this->log( sprintf( '+ ' . __( "Purging post id '%1\$d' ( post type '%2\$s' )", 'nginx-helper' ), $p->ID, $p->post_type ) );
+				$this->log( sprintf( '+ ' . __( "Purging post id '%1\$d' ( post type '%2\$s' )", 'gridpane-nginx-helper' ), $p->ID, $p->post_type ) );
 				$this->purge_url( get_permalink( $p->ID ) );
 
 			}
 		} else {
-			$this->log( '- ' . __( 'No posts', 'nginx-helper' ) );
+			$this->log( '- ' . __( 'No posts', 'gridpane-nginx-helper' ) );
 		}
 
 		return true;
@@ -980,7 +1000,7 @@ abstract class Purger {
 	 */
 	private function _purge_all_date_archives() {
 
-		$this->log( __( 'Purging all date-based archives.', 'nginx-helper' ) );
+		$this->log( __( 'Purging all date-based archives.', 'gridpane-nginx-helper' ) );
 
 		$this->_purge_all_daily_archives();
 		$this->_purge_all_monthly_archives();
@@ -997,7 +1017,7 @@ abstract class Purger {
 
 		global $wpdb;
 
-		$this->log( __( 'Purging all daily archives.', 'nginx-helper' ) );
+		$this->log( __( 'Purging all daily archives.', 'gridpane-nginx-helper' ) );
 
 		$_query_daily_archives = $wpdb->prepare(
 			"SELECT YEAR(post_date) AS %s, MONTH(post_date) AS %s, DAYOFMONTH(post_date) AS %s, count(ID) as posts
@@ -1020,7 +1040,7 @@ abstract class Purger {
 
 				$this->log(
 					sprintf(
-						'+ ' . __( "Purging daily archive '%1\$s/%2\$s/%3\$s'", 'nginx-helper' ),
+						'+ ' . __( "Purging daily archive '%1\$s/%2\$s/%3\$s'", 'gridpane-nginx-helper' ),
 						$_da->year,
 						$_da->month,
 						$_da->dayofmonth
@@ -1031,7 +1051,7 @@ abstract class Purger {
 
 			}
 		} else {
-			$this->log( '- ' . __( 'No daily archives', 'nginx-helper' ) );
+			$this->log( '- ' . __( 'No daily archives', 'gridpane-nginx-helper' ) );
 		}
 
 	}
@@ -1043,7 +1063,7 @@ abstract class Purger {
 
 		global $wpdb;
 
-		$this->log( __( 'Purging all monthly archives.', 'nginx-helper' ) );
+		$this->log( __( 'Purging all monthly archives.', 'gridpane-nginx-helper' ) );
 
 		$_monthly_archives = wp_cache_get( 'nginx_helper_monthly_archives', 'nginx_helper' );
 
@@ -1071,12 +1091,12 @@ abstract class Purger {
 
 			foreach ( $_monthly_archives as $_ma ) {
 
-				$this->log( sprintf( '+ ' . __( "Purging monthly archive '%1\$s/%2\$s'", 'nginx-helper' ), $_ma->year, $_ma->month ) );
+				$this->log( sprintf( '+ ' . __( "Purging monthly archive '%1\$s/%2\$s'", 'gridpane-nginx-helper' ), $_ma->year, $_ma->month ) );
 				$this->purge_url( get_month_link( $_ma->year, $_ma->month ) );
 
 			}
 		} else {
-			$this->log( '- ' . __( 'No monthly archives', 'nginx-helper' ) );
+			$this->log( '- ' . __( 'No monthly archives', 'gridpane-nginx-helper' ) );
 		}
 
 	}
@@ -1088,7 +1108,7 @@ abstract class Purger {
 
 		global $wpdb;
 
-		$this->log( __( 'Purging all yearly archives.', 'nginx-helper' ) );
+		$this->log( __( 'Purging all yearly archives.', 'gridpane-nginx-helper' ) );
 
 		$_yearly_archives = wp_cache_get( 'nginx_helper_yearly_archives', 'nginx_helper' );
 
@@ -1116,12 +1136,12 @@ abstract class Purger {
 			foreach ( $_yearly_archives as $_ya ) {
 
 				// translators: %s: Year to purge cache.
-				$this->log( sprintf( '+ ' . esc_html__( "Purging yearly archive '%s'", 'nginx-helper' ), $_ya->year ) );
+				$this->log( sprintf( '+ ' . esc_html__( "Purging yearly archive '%s'", 'gridpane-nginx-helper' ), $_ya->year ) );
 				$this->purge_url( get_year_link( $_ya->year ) );
 
 			}
 		} else {
-			$this->log( '- ' . __( 'No yearly archives', 'nginx-helper' ) );
+			$this->log( '- ' . __( 'No yearly archives', 'gridpane-nginx-helper' ) );
 		}
 
 	}
@@ -1133,13 +1153,12 @@ abstract class Purger {
 	 */
 	public function purge_them_all() {
 
-		$this->log( __( "Let's purge everything!", 'nginx-helper' ) );
+		$this->log( __( "Let's purge everything!", 'gridpane-nginx-helper' ) );
 		$this->_purge_homepage();
-		$this->_purge_personal_urls();
 		$this->_purge_all_posts();
 		$this->_purge_all_taxonomies();
 		$this->_purge_all_date_archives();
-		$this->log( __( 'Everything purged!', 'nginx-helper' ) );
+		$this->log( __( 'Everything purged!', 'gridpane-nginx-helper' ) );
 
 		return true;
 
@@ -1162,18 +1181,18 @@ abstract class Purger {
 			return;
 		}
 
-		$this->log( __( 'Term taxonomy edited or deleted', 'nginx-helper' ) );
+		$this->log( __( 'Term taxonomy edited or deleted', 'gridpane-nginx-helper' ) );
 
 		$term           = get_term( $term_id, $taxon );
 		$current_filter = current_filter();
 
 		if ( 'edit_term' === $current_filter && ! is_wp_error( $term ) && ! empty( $term ) ) {
 
-			$this->log( sprintf( __( "Term taxonomy '%1\$s' edited, (tt_id '%2\$d', term_id '%3\$d', taxonomy '%4\$s')", 'nginx-helper' ), $term->name, $tt_id, $term_id, $taxon ) );
+			$this->log( sprintf( __( "Term taxonomy '%1\$s' edited, (tt_id '%2\$d', term_id '%3\$d', taxonomy '%4\$s')", 'gridpane-nginx-helper' ), $term->name, $tt_id, $term_id, $taxon ) );
 
 		} elseif ( 'delete_term' === $current_filter ) {
 
-			$this->log( sprintf( __( "A term taxonomy has been deleted from taxonomy '%1\$s', (tt_id '%2\$d', term_id '%3\$d')", 'nginx-helper' ), $taxon, $term_id, $tt_id ) );
+			$this->log( sprintf( __( "A term taxonomy has been deleted from taxonomy '%1\$s', (tt_id '%2\$d', term_id '%3\$d')", 'gridpane-nginx-helper' ), $taxon, $term_id, $tt_id ) );
 
 		}
 
@@ -1201,7 +1220,7 @@ abstract class Purger {
 		switch ( $action ) {
 
 			case 'save-sidebar-widgets':
-				$this->log( __( 'Widget saved, moved or removed in a sidebar', 'nginx-helper' ) );
+				$this->log( __( 'Widget saved, moved or removed in a sidebar', 'gridpane-nginx-helper' ) );
 				$this->_purge_homepage();
 				break;
